Reconfigured file to complete in one call
I rearranged everything so that request handler automatically calls the PGCR parser which in turn uses the CalculateTeamTotals class to generate the reformatted PGCR. From there it is sent back to the request handler. This runs whenever a new instance of request handler is created so that as soon as the request is made it is processed. Will add comments to code shortly -Justin